Backend/pyproject.toml
[tool.poetry]
name = "emsug"
version = "0.0.0"
description = "a emergency management system using genai"
authors = ["Muhammad Fiaz <contact@muhammadfiaz.com>"]
repository = "https://github.com/muhammad-fiaz/Emergency-Management-System-using-Generative-AI.git"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.8.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
